backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-openai
openai
python-dotenv
motor
pinecone-client
redis
pytest
pytest-cov